pyproject.toml
[tool.poetry]
name = "evilflowers-lcpencrypt-worker"
version = "0.1.0"
description = "EvilFlowers Redium lcpencrypt Worker"
authors = ["Jakub Dubec <jakub.dubec@stuba.sk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
celery = {version = "^5.4.0", extras = ["redis"]}
opentelemetry-api = {version = "*", optional = true}
opentelemetry-sdk = {version = "*", optional = true}
opentelemetry-exporter-otlp = {version = "*", optional = true}
opentelemetry-instrumentation-celery = {version = "*", optional = true}

[tool.poetry.extras]
observability = [
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-exporter-otlp",
    "opentelemetry-instrumentation-celery"
]

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

[tool.black]
line-length = 119
target-version = ['py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.github
  | .venv
  | data
)/
'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
